+++ packages/binary/messagelayer/tangle/messagemetadata.go
@@ -10,10 +10,11 @@ import (
 	"github.com/iotaledger/goshimmer/packages/binary/messagelayer/message"
 )
 
+// MessageMetadata defines the metadata for a message.
 type MessageMetadata struct {
 	objectstorage.StorableObjectFlags
 
-	messageId          message.Id
+	messageID          message.ID
 	receivedTime       time.Time
 	solid              bool
 	solidificationTime time.Time
@@ -22,14 +23,16 @@ type MessageMetadata struct {
 	solidificationTimeMutex sync.RWMutex
 }
 
-func NewMessageMetadata(messageId message.Id) *MessageMetadata {
+// NewMessageMetadata creates a new MessageMetadata from the specified messageID.
+func NewMessageMetadata(messageID message.ID) *MessageMetadata {
 	return &MessageMetadata{
-		messageId:    messageId,
+		messageID:    messageID,
 		receivedTime: time.Now(),
 	}
 }
 
-func MessageMetadataFromBytes(bytes []byte) (result *MessageMetadata, err error, consumedBytes int) {
+// MessageMetadataFromBytes unmarshals the given bytes into a MessageMetadata.
+func MessageMetadataFromBytes(bytes []byte) (result *MessageMetadata, consumedBytes int, err error) {
 	marshalUtil := marshalutil.New(bytes)
 	result, err = ParseMessageMetadata(marshalUtil)
 	consumedBytes = marshalUtil.ReadOffset()
@@ -37,17 +40,18 @@ func MessageMetadataFromBytes(bytes []byte) (result *MessageMetadata, err error,
 	return
 }
 
+// ParseMessageMetadata parses the marshalUtil into a MessageMetadata.
+// If it successfully parses the marshalUtil, it delegates to MessageMetadataFromStorageKey.
+// Else, delegates to UnmarshalObjectStorageValue.
 func ParseMessageMetadata(marshalUtil *marshalutil.MarshalUtil) (result *MessageMetadata, err error) {
-	if parsedObject, parseErr := marshalUtil.Parse(func(data []byte) (interface{}, int, error) {
+	parsedObject, parseErr := marshalUtil.Parse(func(data []byte) (interface{}, int, error) {
 		return MessageMetadataFromStorageKey(data)
-	}); parseErr != nil {
+	})
+	if parseErr != nil {
 		err = parseErr
-
 		return
-	} else {
-		result = parsedObject.(*MessageMetadata)
 	}
-
+	result = parsedObject.(*MessageMetadata)
 	_, err = marshalUtil.Parse(func(data []byte) (parseResult interface{}, parsedBytes int, parseErr error) {
 		parsedBytes, parseErr = result.UnmarshalObjectStorageValue(data)
 
@@ -57,11 +61,12 @@ func ParseMessageMetadata(marshalUtil *marshalutil.MarshalUtil) (result *Message
 	return
 }
 
+// MessageMetadataFromStorageKey unmarshals the stored bytes into a MessageMetadata.
 func MessageMetadataFromStorageKey(key []byte) (result objectstorage.StorableObject, consumedBytes int, err error) {
 	result = &MessageMetadata{}
 
 	marshalUtil := marshalutil.New(key)
-	result.(*MessageMetadata).messageId, err = message.ParseId(marshalUtil)
+	result.(*MessageMetadata).messageID, err = message.ParseID(marshalUtil)
 	if err != nil {
 		return
 	}
@@ -75,6 +80,7 @@ func (messageMetadata *MessageMetadata) ReceivedTime() time.Time {
 	return messageMetadata.receivedTime
 }
 
+// IsSolid returns true if the message represented by this metadata is solid. False otherwise.
 func (messageMetadata *MessageMetadata) IsSolid() (result bool) {
 	messageMetadata.solidMutex.RLock()
 	result = messageMetadata.solid
@@ -83,6 +89,8 @@ func (messageMetadata *MessageMetadata) IsSolid() (result bool) {
 	return
 }
 
+// SetSolid sets the message associated with this metadata as solid.
+// It returns true if the solid status is modified. False otherwise.
 func (messageMetadata *MessageMetadata) SetSolid(solid bool) (modified bool) {
 	messageMetadata.solidMutex.RLock()
 	if messageMetadata.solid != solid {
@@ -118,10 +126,14 @@ func (messageMetadata *MessageMetadata) SolidificationTime() time.Time {
 	return messageMetadata.solidificationTime
 }
 
+// ObjectStorageKey returns the key of the stored message metadata object.
+// This returns the bytes of the messageID.
 func (messageMetadata *MessageMetadata) ObjectStorageKey() []byte {
-	return messageMetadata.messageId.Bytes()
+	return messageMetadata.messageID.Bytes()
 }
 
+// ObjectStorageValue returns the value of the stored message metadata object.
+// This includes the receivedTime, solidificationTime and solid status.
 func (messageMetadata *MessageMetadata) ObjectStorageValue() []byte {
 	return marshalutil.New().
 		WriteTime(messageMetadata.ReceivedTime()).
@@ -130,6 +142,7 @@ func (messageMetadata *MessageMetadata) ObjectStorageValue() []byte {
 		Bytes()
 }
 
+// UnmarshalObjectStorageValue unmarshals the stored bytes into a messageMetadata.
 func (messageMetadata *MessageMetadata) UnmarshalObjectStorageValue(data []byte) (consumedBytes int, err error) {
 	marshalUtil := marshalutil.New(data)
 
@@ -148,28 +161,34 @@ func (messageMetadata *MessageMetadata) UnmarshalObjectStorageValue(data []byte)
 	return
 }
 
+// Update updates the message metadata.
+// This should never happen and will panic if attempted.
 func (messageMetadata *MessageMetadata) Update(other objectstorage.StorableObject) {
 	panic("updates disabled")
 }
 
 var _ objectstorage.StorableObject = &MessageMetadata{}
 
+// CachedMessageMetadata is a wrapper for stored cached object that represents a message metadata.
 type CachedMessageMetadata struct {
 	objectstorage.CachedObject
 }
 
+// Retain registers a new consumer for the cached message metadata.
 func (cachedMessageMetadata *CachedMessageMetadata) Retain() *CachedMessageMetadata {
 	return &CachedMessageMetadata{cachedMessageMetadata.CachedObject.Retain()}
 }
 
+// Unwrap returns the underlying stored message metadata wrapped by the CachedMessageMetadata.
+// If the stored object cannot be cast to MessageMetadata or is deleted, it returns nil.
 func (cachedMessageMetadata *CachedMessageMetadata) Unwrap() *MessageMetadata {
-	if untypedObject := cachedMessageMetadata.Get(); untypedObject == nil {
+	untypedObject := cachedMessageMetadata.Get()
+	if untypedObject == nil {
 		return nil
-	} else {
-		if typedObject := untypedObject.(*MessageMetadata); typedObject == nil || typedObject.IsDeleted() {
-			return nil
-		} else {
-			return typedObject
-		}
 	}
+	typedObject := untypedObject.(*MessageMetadata)
+	if typedObject == nil || typedObject.IsDeleted() {
+		return nil
+	}
+	return typedObject
 }