+++ plugins/webapi/drng/collectivebeacon/handler.go
@@ -3,7 +3,7 @@ package collectivebeacon
 import (
 	"net/http"
 
-	"github.com/iotaledger/goshimmer/packages/binary/drng/subtypes/collectiveBeacon/payload"
+	"github.com/iotaledger/goshimmer/packages/binary/drng/subtypes/collectivebeacon/payload"
 	"github.com/iotaledger/goshimmer/plugins/issuer"
 	"github.com/iotaledger/hive.go/marshalutil"
 	"github.com/labstack/echo"
@@ -28,7 +28,7 @@ func Handler(c echo.Context) error {
 	if err != nil {
 		return c.JSON(http.StatusBadRequest, Response{Error: err.Error()})
 	}
-	return c.JSON(http.StatusOK, Response{ID: msg.Id().String()})
+	return c.JSON(http.StatusOK, Response{ID: msg.ID().String()})
 }
 
 // Response is the HTTP response from broadcasting a collective beacon message.